Made a new modifier key, '*', which enters "box selection hold" mode so that you can quickly fill in multiple cells in the same 3x3 box
This mode can be entered by pressing the * key before or after selecting a box and filling in a cell (besides at the very start of a puzzle, when no boxes/cells have been selected or filled). The last selected box is always stored, so when * is pressed the last selection is "held", so the next keypress after a cell is filled selects the next cell to fill within the same box, rather than selecting a box again.